dynamic config based on url.
