+++ src/components/ConfigurationUI.js
@@ -10,7 +10,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import ClearIcon from "@material-ui/icons/Clear";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import {default_configs} from "../Config"
-import { TabbedShowLayout, Tab } from 'react-admin';
+import { TabbedShowLayout, Tab, useRefresh } from 'react-admin';
 import Box from '@mui/material/Box';
 import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
@@ -20,13 +20,20 @@ import Modal from '@material-ui/core/Modal';
 import Typography from '@material-ui/core/Typography';
 import {useNotify } from 'react-admin';
 import {Loading } from 'react-admin';
-
+import  LoadYaml   from './LoadYaml';
 const yaml = require('js-yaml')
 
-let als_yaml_url = "/ui/admin/admin.yaml"
+const str = window.location.href
+const arr=str.split("/")
+const index = arr.findIndex(e=>e==="admin")
+const yamlName = arr[index+1]??"admin"
+
+let als_yaml_url = `/ui/admin/${yamlName}.yaml`
 if(window.location.href.includes(":3000")){
-    als_yaml_url = "http://localhost:5656/ui/admin/admin.yaml"
+    als_yaml_url = `http://localhost:5656/ui/admin/${yamlName}.yaml`
 }
+
+
 
 const useStyles = makeStyles((theme) => ({
     widget : {
@@ -71,72 +78,10 @@ const DeleteConf = (conf_name) => {
 }
 
 
-const addConf = (conf) => {
-    console.log(conf)
-    const configs = JSON.parse(localStorage.getItem("raconfigs"));
-    if(!conf.api_root){
-        console.warn("Config has no api_root", conf);
-        return false
-    }
-    configs[conf.api_root] = conf
-    localStorage.setItem("raconf", JSON.stringify(conf));
-    localStorage.setItem("raconfigs", JSON.stringify(configs));
-    window.location.reload();
-    return true
-}
 
 
-export const LoadYaml = (config_url, notify) => {
-    
-    if(config_url == null){
-        config_url = als_yaml_url
-    }
-    
-    const saveConf = (conf_str) => {
-        // first try to parse as json, if this doesn't work, try yaml
-        try{
-            const conf = JSON.parse(conf_str)
-            if(typeof r !== 'object'){
-                saveYaml(conf_str)
-                return
-            }
-            if( ! addConf(conf) && notify){
-                notify('Failed to load config', 'warning')
-            }
-        }
-        catch(e){
-            saveYaml(conf_str)
-        }
-    }
 
-    const saveYaml = (ystr) => {
-        
-        try{
-            const conf = yaml.load(ystr)
-            if( ! addConf(conf) && notify){
-                notify('Failed to load config', 'warning')
-            }
-        }
-        catch(e){
-            console.warn(`Failed to load yaml`, ystr)
-            console.error(e)
-        }
-    }
 
-    fetch(config_url, {cache: "no-store"})
-        .then((response) => response.text())
-        .then((conf_str) => {
-            saveConf(conf_str)
-            notify("Loaded configuration")
-        })
-        .catch((err)=>{
-                if(notify){
-                        notify("Failed to load yaml", { type : 'warning' })
-                }
-                console.error(`Failed to load yaml from ${config_url}: ${err}`)
-        })
-    
-}
 
 
 const ManageModal = () => {
@@ -279,6 +224,7 @@ export const resetConf = (notify) => {
     LoadYaml(als_yaml_url, notify)
     return defconf
 }
+
 
 const ConfigurationUI = () => {
